+++ packages/shared/src/index.ts
@@ -13,3 +13,4 @@ export * from "./lpEnumeration.js";
 export * from "./dust.js";
 export * from "./b20.js";
 export * from "./solana.js";
+export * from "./cctp.js";